Add windows builds (#192)
Add windows build and fix all related build failures. This currently only builds
`msvc-15` due to LLVM issues. Reduced build size by removing incremental linking
option, switching submodules to do shallow clones and removing unneeded LLVM
targets.

+++ Overlays/Vulkan/Runtime/WNGraphics/src/Vulkan/WNDevice.cpp
@@ -1847,7 +1847,7 @@ bool vulkan_device::initialize_buffer(
 }
 
 bool vulkan_device::bind_buffer(
-    buffer* _buffer, arena* _arena, const size_t _offset) {
+    buffer* _buffer, arena* _arena, const uint64_t _offset) {
   memory::unique_ptr<buffer_info>& bdata = get_data(_buffer);
   const arena_data& adata = get_data(_arena);
