+++ Runtime/WNGraphics/inc/WNBuffer.h
@@ -67,7 +67,7 @@ public:
     return *this;
   }
 
-  WN_FORCE_INLINE bool bind_memory(arena* _arena, const size_t _offset) {
+  WN_FORCE_INLINE bool bind_memory(arena* _arena, const uint64_t _offset) {
     WN_DEBUG_ASSERT(_offset + size() <= _arena->size(),
         "binding is out of bounds of arena");
